feature(exo): move loading spinner to redux
