+++ src/actions/index.js
@@ -1,9 +1,16 @@
 export const Types = {
-  SetListings: 'SET_LISTINGS',
+  SET_LISTINGS: 'SET_LISTINGS',
+  LOADING: 'LOADING'
 };
 export const Actions = {
   setListings: results => ({
     type: Types.SetListings,
     payload: results,
   }),
+  loading: (isLoading) => ({
+    type: Types.LOADING,
+    payload: {
+      isLoading
+    }
+  })
 };
